requirements.txt
kafka-python
python-dotenv
rsa
Flask